Update using current main files
