+++ include/input.h
@@ -5,10 +5,12 @@
 #include<stdlib.h>
 #include<stdio.h>
 #include<ctype.h>
+#include "../include/day.h"
 
 char input_char( char letter); // handle yes no input
 int input_dec_number(int number); //handle the input for number of days or miles
-double input_time(double time); // 00:00 - 23:59: valid time
+void input_time(Day * day); // 00:00 - 23:59: valid time
+int validTime(int hours, int min);
 
 // Andrew's part
 double input_fee(double minimum); // retrieve user fee input
